requirements.txt
# SPDX-FileCopyrightText: 2024 Davíð Berman <davidjberman@gmail.com>
#
# SPDX-License-Identifier: AGPL-3.0-or-later

Flask
requests
python-dotenv
pytz
ephem
